Move default translations from espoo/ to defaults/
