+++ frontend/src/lib-customizations/defaults/employee/i18n/vasu/sv.ts
